# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.AuthResponseData:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.UserRecord'
    type: object
...
  models.PasswordLoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  models.RegisterRequestOTPInput:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
  models.RegisterVerifyOTPInput:
    properties:
      email:
        type: string
      name:
        type: string
      otpCode:
        type: string
      password:
        type: string
    type: object
...
  models.UserRecord:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      isVerified:
        type: boolean
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:8082
info:
  contact:
    email: support@symphoniatic.id
    name: SymphoniaTic API Support
  description: Interactive OpenAPI/Swagger UI documentation for SymphoniaTic Concert
    Ticketing Platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SymphoniaTic REST API Documentation
  version: "1.0"
paths:
...
  /auth/login/password:
    post:
      consumes:
      - application/json
      description: Autentikasi user menggunakan email dan password tradisional.
      parameters:
      - description: Login Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.PasswordLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: Login berhasil!
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.AuthResponseData'
              type: object
        "401":
          description: Email atau password salah
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Login dengan Email & Password
      tags:
      - Auth - User
  /auth/register/request-otp:
    post:
      consumes:
      - application/json
      description: Mengirimkan 6 digit kode OTP registrasi ke email pengguna.
      parameters:
      - description: Register Request OTP Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequestOTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: Kode OTP registrasi berhasil dikirim
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Payload tidak valid / Email terdaftar
          schema:
            $ref: '#/definitions/models.APIResponse'
        "429":
          description: Cooldown OTP (tunggu 60s)
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Meminta kode OTP untuk registrasi
      tags:
      - Auth - User
  /auth/register/verify-otp:
    post:
      consumes:
      - application/json
      description: Memverifikasi kode OTP registrasi dan membuat akun baru beserta
        JWT token.
      parameters:
      - description: Register Verify OTP Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.RegisterVerifyOTPInput'
      produces:
      - application/json
      responses:
        "201":
          description: Registrasi akun berhasil!
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.AuthResponseData'
              type: object
        "400":
          description: OTP tidak sesuai / expired
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Verifikasi OTP & Buat Akun Pengguna
      tags:
      - Auth - User
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your_token>" to authenticate user/admin endpoints.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"